Move client requests tests into http_requests/
