+++ client/verta/tests/http_requests/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Awkwardly named to avoid conflict with the requests library."""
